feat: change weights from double to int16_t and potential and time values from double to float

+++ include/LIF_neuron.h
@@ -1,7 +1,7 @@
 #pragma once
 ////////////////////// LIFNeuron struct //////////////////////////////
 typedef struct {
-    double mp; // membrane potential
-    double ts; // time-stamp of last spike
+    float mp; // membrane potential
+    float ts; // time-stamp of last spike
 } LIFNeuron;
 //////////////////////////////////////////////////////////////////////
